refactor: decorator types
